% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneinfo.R
\name{getGeneinfo}
\alias{getGeneinfo}
\title{Get Gene Info}
\usage{
getGeneinfo(ensembl_id)
}
\arguments{
\item{ensembl_id}{An Ensemble ID}
}
\value{
A dataframe with gene info details
}
\description{
Perform lookups of Ensembl Identifiers and retrieve their 
external references in other databases using \href{http://rest.ensembl.org/documentation/info/xref_id}{xref_id-ENSEMBL API}
}
\examples{
mygeneinfo <- getGeneinfo("ENSG00000157764")
head(mygeneinfo)

}
