% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Make leaflet map widget}
\usage{
make_map(title = "This is a test", lat, lng)
}
\arguments{
\item{title}{label for the marker}

\item{lat}{lattitude coordinate}

\item{lng}{longintude coordinate}
}
\description{
Wrapper for link[leaflet]{leaflet} package. 
\url{http://rstudio.github.io/leaflet/}.
}
