% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(file, ...)
}
\arguments{
\item{file}{Input a .csv or .tsv file}

\item{...}{Further arguments to be passed to \code{\link[pheatmap]{pheatmap}}}
}
\value{
A heatmap
}
\description{
This is a wrapper function to \code{\link[pheatmap]{pheatmap}} taking a file instead of matrix
}
\examples{
file <- system.file("extdata", "pasilla_gene_counts.csv", package = "sahu")
plotHeatmap(file, main='heatmap title')

}
\author{
Sangram Keshari Sahu
}
\keyword{heatmap}
