% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA3d.R
\name{plotPCA3d}
\alias{plotPCA3d}
\title{Plot 3D-PCA}
\usage{
plotPCA3d(file)
}
\arguments{
\item{file}{Input a .csv or .tsv file}
}
\value{
a PCA-plot
}
\description{
This function returns an 3D PCA plot
}
\examples{
file <- system.file("extdata", "expr_matrix.csv", package = "sahu")
plotPCA3d(file)

}
\author{
Sangram Keshari Sahu, \email{sangramsahu15@gmail.com}
}
\keyword{pca-3d}
