% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resave.R
\name{resave}
\alias{resave}
\title{resave}
\usage{
resave(..., list = character(), file)
}
\arguments{
\item{...}{A list of vectors separated by coma(,)}

\item{list}{A charater vector list}

\item{file}{An existing .RData file on which append will be done.}
}
\description{
This is a wrapper to `save()` to append obejcts to existing .RData file. 
https://stackoverflow.com/questions/11813096/updating-an-existing-rdata-file
}
