% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{who}
\alias{who}
\title{Who
This function will tail details about the author}
\usage{
who()
}
\value{
this returns name
}
\description{
Who
This function will tail details about the author
}
\examples{
who()

}
\references{
\url{https://sksahu.net}
}
\seealso{
\code{\link{who}}
}
\author{
Sangram, \email{sangramsahu15@gmail.com}
}
\keyword{function}
\keyword{initial}
